Reduced memory count in client
Now only value id is tracked for received instances. This is
accomplished by introducing text based header at the start of each
value, which includes value id and md5 checksum for value itself as
well as value id.

This is enough for verification that received value is correct without
extra copy on the client side, so client now only tracks value id and
full value only for instances that have not been received yet and thus
might require resending.