Fixed zero ballot numbers
New WAL format implementation has check upon reading a record that b
and vb are positive non-null integers, and values which were delivered
by the learner were saved with b = 0. I changed it so that b = vb now.

It should not break the protocol (and it's implementation), because as
far as I remember b is unused for records that are finalized already.