Dropped COW semantics for acceptor storage instances
This should not affect the safety. Snapshots must not be consistent
among several instances as they are independent from each other. OTOH
snapshots must be consistent about each single instance and
corresponding place in LSN sequence. Without COW some instance can be
flushed to the snapshot while being dirty, which means that it has not
yet found its way to the WAL and responses were not send about it. LSN
of the snapshot will be the one currently confirmed, which means that
all dirty records that were put into this snapshot will be overwritten
after they are flushed into the next WAL. In case a process fails after
writing the snapshot, but before flushing the WAL and sending the
responses we may just assume that the messages were lost, which does
not affect Paxos safety.

With this change it might be more dangerous to clone a snapshot from one
node to another without some trailing WAL logs, but that process is yet
to be figured out and documented.